+++ sdk/core/azure-core-http-okhttp/src/main/java/com/azure/core/http/okhttp/OkHttpAsyncClientProvider.java
@@ -33,7 +33,7 @@ public final class OkHttpAsyncClientProvider implements HttpClientProvider {
     @Override
     public HttpClient createInstance() {
         final String disableDefaultSharingHttpClient =
-            Configuration.getGlobalConfiguration().get("AZURE_DISABLE_DEFAULT_SHARING_HTTP_CLIENT");
+            Configuration.getGlobalConfiguration().get("AZURE_DISABLE_HTTP_CLIENT_SHARING");
         if ("true".equalsIgnoreCase(disableDefaultSharingHttpClient)) {
             return new OkHttpAsyncHttpClientBuilder().build();
         }