+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/models/CloudServiceExtensionProfile.java
@@ -5,16 +5,12 @@
 package com.azure.resourcemanager.compute.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 /** Describes a cloud service extension profile. */
 @Fluent
 public final class CloudServiceExtensionProfile {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(CloudServiceExtensionProfile.class);
-
     /*
      * List of extensions for the cloud service.
      */